Added Log Utils & Foreground Service for data transfer

+++ app/src/main/java/com/example/obdscanner/model/dto/SensorReading.kt
@@ -9,5 +9,5 @@ data class SensorReading(
     val pid : Byte,
     val name : String,
     val unit : String,
-    var data : Int
+    var data : Double
 )
